requirements.txt
CherryPy==18.5.0
fitbit==0.3.1
Flask==1.1.1
matplotlib==3.2.1
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
polar-accesslink==0.0.5
PyYAML==3.12
requests==2.18.4
scikit-learn==0.22
scipy==1.4.1
